requirements.txt
ema_pytorch
accelerate
einops
tqdm
pymunk
pygame
imageio
shapely